+++ stepfit.h
@@ -40,11 +40,11 @@ struct Data_Struct
     double maxbaseline;
     double risetime;
     int sign;
-    double (*stepfunc)(double time, const double *p, double maxlength, double maxstep, double maxbaseline, double risetime, int sign);
+    double (*stepfunc)(double time, const long double *p, double maxlength, double maxstep, double maxbaseline, double risetime, int sign);
 };
 typedef struct Data_Struct data_struct;
 
-double stepfunc(double time, const double *p, double maxlength, double maxstep, double maxbaseline, double risetime, int sign);
+double stepfunc(double time, const long double *p, double maxlength, double maxstep, double maxbaseline, double risetime, int sign);
 void time_array(double *time, int64_t m);
 void step_response(event *current, double risetime, int64_t maxiters, double minstep);
